feat: Modify keyboard and mouse sharing
Modify keyboard and mouse sharing

Log: Modify keyboard and mouse sharing

+++ src/plugins/daemon/core/service/searchlight.cpp
@@ -113,7 +113,7 @@ void Discoverer::start()
 //            break;
         }
 
-        fastring msg(buffer, recv_len);
+        fastring msg(buffer, static_cast<size_t>(recv_len));
         handle_message(msg, co::addr2str(&cli));
     }
